Data base changed/ don't working now
