+++ src/components/card_modal/index.tsx
@@ -1,0 +1,2 @@
+import CardModal from "./card_modal";
+export default CardModal;
