--- src/components/card_sticker/index.tsx
@@ -1,2 +1,0 @@
-import CardSticker from "./card__sticker";
-export default CardSticker;
