+++ src/settings.tsx
@@ -15,7 +15,13 @@ export const settings = {
         comments: "Комментарии",
         text: "Описание",
         creator: "Создал:",
-
+        emptyCard: {
+            comments: [],
+            id: 0,
+            name: "",
+            text: "",
+            user: "",
+        },
     },
     comments:{
         name: "Комментарии",
@@ -37,6 +43,38 @@ export const settings = {
     },
     welcome:{
         placeholder: "Введите ваше имя здесь.",
+    },
+    baseName: "t",
+    emptyData:  {
+        columns: [
+            {
+                id: 1,
+                title: "TODO",
+                cards: []
+            },
+            {
+                id: 2,
+                title: "IN PROGRESS",
+                cards: []
+            },
+            {
+                id: 3,
+                title: "TESTING",
+                cards: []
+            },
+            {
+                id: 4,
+                title: "DONE",
+                cards: []
+            }
+        ]
+
+        // column: {
+        //     1: {title: "TODO", content: []},
+        //     2: {title: "IN PROGRESS", content: []},
+        //     3: {title: "TESTING", content: []},
+        //     4: {title: "DONE", content: []},
+        // },
     }
 
 }
