+++ src/types/index.tsx
@@ -1,0 +1,3 @@
+import {ColumnsType, CardType, CommentType, DataType} from "./types";
+
+export type {DataType, CommentType, CardType, ColumnsType}
